Convert myth timers to std::chrono. (1)
This commit focuses on converting MythTimer, and propagating those
changes throughout the rest of the code without changing too many
other interfaces. There are some temporary conversions to/from
int/std::chrono::millisecond that should be removed with later
commits.

+++ mythtv/libs/libmythtv/tvbrowsehelper.cpp
@@ -147,7 +147,7 @@ void TVBrowseHelper::BrowseDispInfo(const BrowseInfo& Browseinfo)
         return;
 
     m_parent->KillTimer(m_browseTimerId);
-    m_browseTimerId = m_parent->StartTimer(static_cast<int>(TV::kBrowseTimeout), __LINE__);
+    m_browseTimerId = m_parent->StartTimer(TV::kBrowseTimeout, __LINE__);
 
     QMutexLocker locker(&m_browseLock);
     if (BROWSE_SAME == Browseinfo.m_dir)
